Initial reference implementation
Created a reference implementation for a multiset based on Map[A, Int].
This will only be used to test the more advanced implementations and
shouldn't be used by the end-user.

I still need to sit down a come up with a proper well thought out API
for Multisets but right now I just wanted to get some coding done.