Tied primitive implementation into the collections
I did this mostly to experiment with extending the Scala collection.

I made my primitive implementation extend IterableLike. This meant
that I had to implement empty, iterator, newBuilder, and add the
correct `implicit def canBuildFrom` to make `map` and friends
return a new ListMultiset when possible.

I might have to tweak the implicit a bit as it currently only
returns a new ListMultiset when the functions are A => A which
shouldn't be a requirement.