Added a multiset that doesn't merge values
We now have two primitive implementations of
immutable multisets. Also introduced a
trait (type-class) for things that can be
merged called `Mergeable`. The goal is to
make the apply of immutable.Multiset select
the most appropriate Multiset (i.e. select
a multiset that keeps all values if the
value types aren't `Mergeable`).