Removed trait Mergeable
I initially wanted the trait Mergeable[A] as a flag to
the compiler so that it knew when it was okay to use a
Multiset implementation that merged equal values
together and stored a counter instead. However, with the
hierarchy I've created with a clear distinction between
Multiset and MergeableMultiset this seems to just add a
bit of unnecessary boilerplate, for example, if you have
a MergeableMultiset[A] and want to map
mset.map(_.toString) then you also need to have an
`implicit val Mergeable[String]` in scope for it to work
and the error you get is quite subtle (you get a
TraversableOnce[String] instead of a
MergeableMultiset[String] without an explanation why).
So as it is currently I don't think it adds any value
but it can get in the way.